Run workflow on pull requests (#11)
As part of the new branch protection rule to require status checks to
pass before merging, we need to run the `build-and-test` job before a
pull request can be closed for `main`. This PR modifies the workflow to
run on this scenario.

- Adds a trigger so that the workflow runs on pull request events, the
run status will now show on the pull request page when it completes.
- Adds a condition to the `deploy` job so that it only runs in the
`main` branch. i.e. when the `main` branch has been updated. This
prevents the code from being deployed from the pull request itself
before its merged, for example.