Deduct tax withheld from budget (#35)
Creates a new field that sums the invoice total plus the tax withheld
called "total spent amount". This new field will be deducted from the
budget since it represents the total costs for that invoice. i.e. the
total paid plus the taxes that will need to be paid later for that
invoice.