Remove placeholder text from forms (#76)
As part of an initial design idea, I left some placeholder text to add
an informational alert to the side of the form. I decided to leave this
information, if any, at the top of the form instead and allow the form
contents to occupy the full container width.

As a first step to remove this two-column layout for forms, I am
removing the placeholder text.

This is the first PR to complete #75