Fix Result Create cancel navigation (#80)
Fixes an issue where the user would be redirected to a non-existing page
when cancelling the Result creation form.

The problem was introduced when the query parameter that contains the
current page Project ID was changed from `projectId` to `project`. This
parameter is needed to create a link to the appropriate Project Budget
page to go "back" to.

This PR fixes the query parameter name to correctly extract the current
project the Result is being created to, so the user returns to the
correct Project Budget page.