Add Project-level reports (#81)
Currently the application only generates reports of the entire database
(organization-wide data). With this PR, the functionality is added to
create reports for only the selected Project. Allowing the user to
choose the scope of the reports and use the database resources more
intelligently.

- Adds the new reports to the Reports dashboard along with the localized
strings.
- Adds the required logic to filter the database queries per project, by
extending the existing report generation infrastructure.
- Adds the logic to display a select list of projects in the Report
dashboard.
- Minor string adjustments.

Closes #73